weaken schema check on mem table
